requirements.txt
flask
flask_restful
bcrypt
mysql-connector-python
gunicorn
unittest2